requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Babel==2.14.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
mysqlclient==2.1.1
pycodestyle==2.10.0
pyindia-zipcode==0.0.2
pytz==2023.3.post1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2022.7
urllib3==1.26.15
